requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
httpx
requests
cryptography
python-jose[cryptography]
passlib[bcrypt]
openpyxl
python-multipart
